+++ backend/tests/Unit/Slim/SessionMiddlewareTest.php
@@ -6,14 +6,9 @@ namespace Tests\Unit\Slim;
 
 use Neucore\Slim\SessionMiddleware;
 use Neucore\Service\SessionData;
-use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\ResponseInterface;
-use Slim\Interfaces\RouteInterface;
-use Slim\Interfaces\RouteParserInterface;
-use Slim\Routing\RouteContext;
-use Slim\Routing\RoutingResults;
-use Tests\RequestFactory;
 use Tests\RequestHandler;
+use Tests\Unit\TestCase;
 
 class SessionMiddlewareTest extends TestCase
 {
@@ -88,19 +83,7 @@ class SessionMiddlewareTest extends TestCase
 
     private function invokeMiddleware(array $conf, string $path = null): ResponseInterface
     {
-        $routeParser = $this->createMock(RouteParserInterface::class);
-        $routingResults = $this->createMock(RoutingResults::class);
-        $route = $this->createMock(RouteInterface::class);
-
-        $request = RequestFactory::createRequest();
-        $request = $request->withAttribute(RouteContext::ROUTE_PARSER, $routeParser);
-        $request = $request->withAttribute(RouteContext::ROUTING_RESULTS, $routingResults);
-
-        if ($path) {
-            /* @phan-suppress-next-line PhanAccessMethodInternal */
-            $route->method('getPattern')->willReturn($path);
-            $request = $request->withAttribute(RouteContext::ROUTE, $route);
-        }
+        $request = $this->createRequestWithRoute('GET', $path);
 
         $nbs = new SessionMiddleware($conf);