+++ backend/tests/Unit/TestCase.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace Tests\Unit;
+
+use Psr\Http\Message\ServerRequestInterface;
+use Slim\Interfaces\RouteInterface;
+use Slim\Interfaces\RouteParserInterface;
+use Slim\Routing\RouteContext;
+use Slim\Routing\RoutingResults;
+use Tests\RequestFactory;
+
+class TestCase extends \PHPUnit\Framework\TestCase
+{
+    protected function createRequestWithRoute(string $method = 'GET', string $path = null): ServerRequestInterface
+    {
+        $routeParser = $this->createMock(RouteParserInterface::class);
+        $routingResults = $this->createMock(RoutingResults::class);
+        $route = $this->createMock(RouteInterface::class);
+
+        $request = RequestFactory::createRequest($method, $path ? $path : '/');
+        $request = $request->withAttribute(RouteContext::ROUTE_PARSER, $routeParser);
+        $request = $request->withAttribute(RouteContext::ROUTING_RESULTS, $routingResults);
+
+        if ($path) {
+            /* @phan-suppress-next-line PhanAccessMethodInternal */
+            $route->method('getPattern')->willReturn($path);
+            $request = $request->withAttribute(RouteContext::ROUTE, $route);
+        }
+
+        return $request;
+    }
+}